requirements.txt
PyQt6>=6.9.0
mysql-connector-python>=9.4.0